Added the SchemaObject.observe_relation_changes() method.
Some models call for computed fields with values derived from related elements
(ie. a field indicating if an hotel is full, depending on wether all of its
rooms have been booked). The usual way to deal with this was to set up a chain
of several event handlers on the related objects. This was cumbersome and error
prone; this method streamlines the whole process and makes it easy for a model
to listen to changes on the fields of related objects.