Changed the implementation for HTML element client models, parameters and code.
Client models posed a problem when added to the cache, since they are not
guaranteed to be serializable - elements have any attribute on them, with any
sort of values and data types - and many cache backends serialize their content.

To work around this, client models are no longer stored away to be rendered
late, when the whole document is rendered. Instead, they are rendered following
the ordinary rendering sequence of any other element, but their output and
metadata is stored on a secondary rendering buffer.

Since this complicated the handling of their client parameters and code, these
two aspects have also been reimplemented. Instead of being relayed through
javascript snippets, they are now passed on as HTML5 data-* attributes of the
element they belong to. This means that is no longer possible to use an HTML4
renderer and client parameters / code and pass validation. On the plus side,
this should make for terser pages, and is a much more elegant and simpler
solution.