Introduced the Cache.retrieve_with_metadata() method.
Some times it is necessary to obtain not only the value that was added to the
cache, but also its expiration and tags. The new method returns a tuple
containing all this data. The REST cache server has been changed accordingly.

The driving use case for this feature is the need to preserve cache tags for
HTML elements retrieved from the cache (see the next changeset for details).