Added the html.Element.update_cache_expiration() method.
This was previously defined by Woost as an extension to the Element class, to be
used by the .depends_on() method (another Woost related extension). But in
reality, the method has nothing to do with with Woost, and can be useful on its
own. Therefore, it has been moved upstream.