Don't apply a stop words filter to full text indexing.
Full text indexes applied a stop words filter provided by zope.index that
filtered some common english words. These presented two problems:

* The filter was applied during indexing, but not to queries. That meant that
AND queries (incidentally, the default) returned 0 matches if they included any
of the words in the stop words filter.

* The filter's list of words was limited to common english words, regardless of
the language/s of the text placed in the index.

To solve these problems, the filter has been removed. This is a bit of a cop
out, since the filter's functionality was desirable. It could be enabled again
in the future, but it should be reimplemented in a way that avoids these
pitfalls.