The CML compiler now strips whitespace from text nodes.
The py:strip_whitespace directive was supposed to do this, but in practice,
stripping should be the default behavior, as it reduces memory consumption and
processing time and most whitespace elements were noise anyway. Also, the
directive was buggy and didn't work in all cases.

That said, this change can potentially affect some templates. Since the negative
effects are relatively mild (at worst, some texts or elements may be strung
together), a new branch was not deemed necessary.