HTML documents should always include the core scripts.
Previously, jQuery and cocktail/scripts/core.js were only included if passing
data or initialization code from the server to the client was required. Since
that was so prevalent, most scripts didn't bother with stating their
dependencies on the core scripts. Hence, including them automatically seems like
a better choice. In the future, a property on the Document class could be added
to make it possible to disable this behavior.