Introduced the py:overlay_base attribute for CML based overlays.
Overlays defined using a CML template were assumed to inherit from
cocktail.html.Overlay. This is fine in most cases, but sometimes it can be
useful to inherit from a subclass of Overlay, to encapsulate common behaviors.

The new py:overlay_base attribute directive allows to indicate to the CML
template compiler which class should be used as an overlay's base.

An example use case of this feature is the per-site customization of templates
applied by palaumusica.org, which makes it possible to have overlays that only
apply to intranet or frontend pages, by using IntranetOverlay / FrontendOverlay
as their bases.