Use DynamicDefault to generate incremental IDs for indexed persistent classes.
Previously, the generation was invoked explicitly by the constructor of the
PersistentObject class, but this seems to be unnecessary. Moreover, the new
possibility of passing schema.undefined to the constructor can help prevent the
useless acquisition of incremental IDs in many cases (f. eg. when serving forms
based on a model).