Optimization for the selectable() client side behavior.
The implementation of the :selectable-entry pseudo element was terribly slow, as
it relied on the :has() selector. Using a "subquery" and an intersection yields
better results.

This change speeds up the opening / closing of dropdown elements with big check
lists (for example, the translation selectors in the woost backoffice for the
Mango / Help site).