Redesigned the containingDropdownExpanded/Collapsed events to fix performance.
Triggering those events on all the items within a DropdownPanel was very
expensive, and could cascade trigger the same event handler over an over (ie. a
SearchableCheckList handles the event, but since it is triggered on every
element, including its entries, bubbling causes its handler to be triggered
multiple times).

The new mechanism relies on consumers of the event adding an "exposable" CSS
class on those elements that care about the event. Also, the events have been
renamed to 'exposed' and 'concealed'.