Changed the signature of the html.Selector.create_entry() method.
Selector subclasses that wanted to override this method to change the display of
their entries received a "value, label, selected" triad. This was enough for
simple cases (such as RadioSelector, DropdownSelector and CheckList), but the
lack of a reference to the item that the entry represented was absolutely
crippling to many other use cases. Many subclasses resorted to reobtaining the
reference through the item's identifier (the "value" portion), which was
inefficient and ugly.

The new signature passes a single parameter to create_entry(): the item that the
entry represents. The value, label and selection state for the entry can be
obtained by those implementations that need them by calling the
get_item_value(), get_item_label() and is_selected() methods.

As a side effect, the get_entry_url() and create_entry_link() methods of
LinkSelector have also changed their signatures to receive a single argument.