Support for reading schema.Mapping members from two different parameters.
Previously, reading a schema.Mapping from a request was only possible if the
member was serialized and contained within a single parameter. With this change,
it is now also possible to read mappings from two separate collection
parameters: one for the keys (with a '-keys' suffix), and one for the values
(with a '-values' suffix).

For example, for a request with the following parameters:

    ?foo-keys=a&foo-keys=b&foo-values=1&foo-values=2

Reading a parameter like this:

    get_parameter(
        schema.Mapping("foo",
            keys = schema.String(),
            values = schema.Integer()
        )
    )

Would result in the following value:

    {"a": 1, "b": 2}