Moved the support for read only fields from html.Form to UIGenerator.
Previously, the decision to use a read only display for a field was done at
html.Form. The problem with this approach is that it makes it hard to inherit
this behavior (for example, when a nested compound field delegates the creation
of its displays to its parent).

The new approach is to move this to the UIGenerator itself, with its new
read_only_ui_generator attribute.