Added the html.ValueDisplay class, as a replacement for html.TranslatedValue.
The TranslatedValue class has some issues when used as a display; since it
renders a value directly, without wrapping it inside an HTML element, it can't
expose CSS classes or attributes to stylesheets and scripts. The new
ValueDisplay class provides the old behavior but wraps the resulting values
using a span.