Added the html.Element.init_data_binding_delegate() method.
This change exposes the data binding delegation machinery through a public
method in html.Element. Although this may enable subclasses to customize the
delegation process, the main drive behind the change is to make it possible to
forward data binding configuration from a single element to multiple delegates.
A typical use case for this need is a simple control that embeds a more complex
replacement control based on javascript.