Member.resolve_constraint() failed when the evaluated value was an object.
When deciding the context that an expression for a dynamic constraint should be
evaluated against, Member.resolve_constraint() invoked
ValidationContext.get_object() on the tip of the validation stack. Typically,
this would ascend towards the root, until an object was found; but if the
stack's tip contained a related object, that would be returned instead.

To fix this, start looking for the context on the parent of the stack's tip.