Replace 'preprocessing' with 'filepublication', support source maps for SASS.
The cocktail.preprocessing API was ill suited to dealing with source maps; for
one, it assumed that source files should always be transformed into their
processed form when served, or that they would always be cached in the same way.

The new file_publication API makes far less assumptions, which makes it possible
to implement different processing and serving patterns for different file types.
For example, it can create CSS and source map files on demand for SASS files.

As a side effect, the serve_file() method no longer relies on CherryPy's
implementation, and it now supports serving file-like objects directly.