Don't use the Froogaloop API in the VimeoPlayer view.
Using it with multiple, dynamically created players on a single page (a typical
enough case on media galleries) was completely broken (Froogaloop treats the
ready state as a shared property for all players, which causes the first video
to play correctly, while any following attempt to create a new player fails).

To work around this, we don't use the Froogaloop wrapper anymore, and instead
send messages to Vimeo's iframe using the Window.postMessage() method.

Side effects:

- The vimeoAPI() and vimeoCommand() methods have been removed
- The player now generates 'playing', 'pause' and 'ended' events, following the
  naming conventions of the HTML5 Media Events spec