Bugfix: the SchemaObject constructor would sometimes ignore translated values.
The SchemaObject constructor calls Schema.initialize() to initialize its
instance with default values or supplied arguments. The problem is initialize()
included the initialization of the SchemaObject.translations member. If that
happened *after* setting one or more translated values, those values would be
lost. Since members are initialized in the order resulting from iterating the
schema's members() mapping, the problem affected only members preceding
"translations" in the mapping iteration sequence (which is unpredictable).

The fix removes SchemaObject.translations from the initialization in the
constructor.