Replaced CSRFProtection.should_apply() with an extensible system.
This changeset rolls back the changes applied by 17828542376d. The
should_apply() method is renamed to should_inject(), to clarify the meaning;
the method also calls a new event, 'deciding_injection', which API clients can
listen to in order to raise a CSRFProtectionExemption exception if the
protection scheme shouldn't be injected into the current request.