Added the TreeView.highlighted_selection property.
The property makes it possible to disable markup production for the active entry
and selected branches on those trees that don't require it. This can lead to
performance improvements in some scenarios. It can also be useful to share a
cached tree between different contexts if the selected entry can be highlighted
client side.