Support for mixing py:def + py:for directives in CML templates.
When defining a part of a template that should be repeated, this was the proper
idiom, until now:

<py:ready>
    <py:new
        py:element="self.create_foo_entry(foo)"
        py:for="foo in self.foo_list"/>
</py:ready>

<div py:def="foo_entry" py:args="foo">
    ...
</div>

This allows other templates (ie. subclasses) to override the behavior of the
entries. It involves a lod of boilerplate, though; with the new syntax, it can be
reduced to the following snippet, which is functionally identical:

<div
    py:def="foo_entry"
    py:for="foo in self.foo_list"/>
    ...
</div>