py:def + py:for shouldn't imply a ready stage deferral.
Mixing py:def and py:for has proven to be a very useful shortcut, but assuming
that all its uses should include a deferral until the 'ready' stage was wrong.
Specially, when nesting several such constructs, adding a <py:ready> instruction
implictly can negatively impact performance (at best) or even alter semantics
(at worst) due to the introduction of a closure.

This is a breaking change, so previous usage of this feature that relied on this
behavior should be updated to include the new py:when="ready" directive:

<div py:def="foobar" py:for="foo in bar"/>

Should become:

<div py:def="foobar" py:for="foo in bar" py:when="ready"/>