cocktail.ui: make Selectable trigger a selectionActivated event
Until now, Selectable only notified implementors of single / double click
activations using a selectionActivated() method. Now, a selectionActivated
event will be triggered as well. The method is still available (it can be more
comfortable to define than messing with event handlers, specially if the
activation behavior has to be overriden).