cocktail.ui: added the cocktail.ui.Items component
It is a substitute for most uses of cocktail.ui.List. It offers the same API and
overall behavior, but it is meant as a building piece for other components: it
has no shadow DOM of its own, nor any default styles. cocktail.ui.List remains,
but it should only be used wherever a full blown component with a regular list
presentation (with dots/dashes) makes sense.