cocktail.ui: added the cocktail.ui.DataDisplay.createDefaultDataBinding() method
Rationale: this allows components to provide a default member, if that makes
sense for their use cases; this frees their users from having to declare it
explicitly each time the component is instantiated in a context that doesn't
provide a member definition (ie. outside a Form).

Example: woost.admin.ui.DescriptiveItems defaults to a Collection of Item. This
makes it possible to create an instance of the component while only specifying
a list of Item instances.