Changed the client side Member.translate() method
Look for translations in a Member.translations property, instead of sourcing
cocktail.ui.translations. This makes it possible to export members and their
translations from the server to the client as a self contained expression; so
far this required a separate block of statements to modify the global
translations catalog.